Add build error parser
The parser will parse soong's 'build.log' files, which follow a
well-defined format. The only ambiguity is in the first line of the
Outputs section, which may or may not include '<column>:'.

Also implement a simple main() to call the new functions to get rid of
rustc's dead code warnings.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>