Add build warning parser
The parser will parse soong's 'verbose.log.gz' files and assumes errors
start with '<path>:<line>:<column>: message' and continue until the next
warning, the next line that start with '[<n>/<N>]' or end-of-file,
whichever comes first.

The parser is quite slow. Some quick benchmark figures obtained by hand:

  tests/data/idmap-errors/verbose.log.gz                     0.085s
  tests/data/easter-egg-errors-kt/verbose.log.gz             0.123s
  tests/data/easter-egg-errors-java/verbose.log.gz           0.226s
  tests/data/idmap-warnings/verbose.log.gz                  18.551s
  tests/data/idmap-both-errors-and-warnings/verbose.log.gz  81.057s

This can probably be improved by changing the implementation to parse the input
line by line, and maintain a state machine (is-parsing-a-warning, is-noise).

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>