Parse warnings line by line
Replace the "scan forward to start of next warning, noise or
end-of-file" regular expression based approach with a more direct
parser. Look at each line in isolation while keeping track of a
half-baked Captures object. Amend the object or add it to the output
based on what the next line contains.

Compared to the previous benchmark run (after introducing lazy_static):

    bench-errors    time:   [3.4058 ms 3.4716 ms 3.5001 ms]
                    change: [-5.2119% -1.7686% +1.7658%] (p = 0.36 > 0.05)
                    No change in performance detected.

    bench-warnings  time:   [999.92 ms 1.0041 s 1.0078 s]
                    change: [-58.068% -57.831% -57.596%] (p = 0.00 < 0.05)
                    Performance has improved.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>