Use Criterion's default benchmark setup
With the code somewhat optimized, benchmarks are no longer expected to
take hours to complete when using the default settings. Remove the
hard-coded limits on benchmark duration and number of iterations:
execution time is still in the range of a couple of minutes, and the
results should be more statistically reliable.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>