Import test data: identical errors with different bodies
Import error.log and verbose.log.gz from build where stat(2) is passed
the wrong number of parameters. The project was cross-compiled for ARM,
x86 and x86_64 and contain the same error repeated for every target.
However, the bodies of the errors are slightly different, as the path to
stat.h depends on the target.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>