Remove duplicate items from output
Sort and remove duplicate items before printing to stdout. Items derive
the usual comparison traits except for PartialEq which is explicitly
implemented. The reason for this is to ignore the Item.body field, which
may differ slightly if cross-compiling the same code for different
targets. An example of this type of error is found in
tests/data/idmap-identical-errors.

Issue: #3
Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>