Complain if no errors found in error.log file
Highlight to the end user if no items could be parsed from a non-empty
errors file: this scenario either means the user passed in an invalid
file, or that the error parser needs to be updated with more patterns.

Technically, error::parse now returns a Result. For consistency, update
warning::parse to also return a Result, even though the warnings parser
never returns Err.

Issue: #11
Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>